+++ src/Options.hpp
@@ -267,12 +267,11 @@ struct Options {
   ValOption<int32_t> bitsReduced{"bits-reduced",
                                  "Bit width of maximum coefficient after reduction when exceeding bits-overflow (0 is "
                                  "unlimited, 1 reduces to cardinalities)",
-                                 limitBit<int, long long>(), "0 =< int",
-                                 [](const int32_t& x) -> bool { return x >= 0; }};
+                                 limitBit<int, int64_t>(), "0 =< int", [](const int32_t& x) -> bool { return x >= 0; }};
   ValOption<int32_t> bitsLearned{
       "bits-learned",
       "Bit width of maximum coefficient for learned constraints (0 is unlimited, 1 reduces to cardinalities)",
-      limitBit<int, long long>(), "0 =< int", [](const int32_t& x) -> bool { return x >= 0; }};
+      limitBit<int, int64_t>(), "0 =< int", [](const int32_t& x) -> bool { return x >= 0; }};
   ValOption<float> optRatio{"opt-ratio", "Ratio of bottom-up optimization time (0 means top-down, 1 fully bottom-up)",
                             0.5, "0 =< float =< 1", [](const double& x) -> bool { return x >= 0 && x <= 1; }};
   BoolOption optCoreguided{"opt-coreguided", "Core-guided bottom up optimization instead of a basic approach", true};
@@ -296,7 +295,7 @@ struct Options {
                           "and may yield UNSAT when no solution within the bounds exists.",
                           true};
   ValOption<double> intDefaultBound{"int-defbound", "Default bound used for unbounded integer variables",
-                                    limitAbs<int, long long>(), "0 < double",
+                                    limitAbs<int, int64_t>(), "0 < double",
                                     [](const double& x) -> bool { return x > 0; }};
   BoolOption pureLits{"inp-purelits", "Propagate pure literals", false};
   ValOption<int32_t> domBreakLim{
